Fixed #18375 -- Ensured F-expressions use existing joins consistently
Previously a .filter(revfk__somecol=F('revfk__anothercol')) created two
joins, one for the F() expression, another for the revfk__somecol
lookup. Fixed this in a way that the revfk__somecol join is first
created, then reused for the F() expression.

This is particularly important if there was multiple lookups in one
.filter() call. In that case then join generated could be kwargs order
dependant, and thus dict ordering dependant.

This commit also added a new DupeMultisF() F-expression subclass so
that upgraders can generate the same queries in 1.5 than they did in
1.4.