Fixed #16759 -- Remove use of __deepcopy__ in qs.clone()
The original problem was that queryset could be really expensive in
certain situations, when the __deepcopy__ went too deep copying Meta
attributes of the models used. To fix this, and other speed and code
cleaness issues this commit removes the use of __deepcopy__ in the
qs, and instead uses .clone() in the internals.