Fixed #18702 -- Removed chunked reads from iter(qs)
The queryset iterator protocol involved six(!) nested iterators. This
was complex to understand and also caused some small performance
impact.

The reason for the complex setup was support for converting the objects
only in small batches to objects. So, one could have multiple iterators
against the same queryset, and all of the iterators would need to make
sure they access the underlying cursor's iterator in synchronized manner
and yet do not interfere with each other.

After the patch the whole queryset is converted to objects in one go,
and after that it is trivial to support concurrent iterators to the
same queryset. The cost of this is always converting all the objects
to Python while the old code was able to convert only a part of the
cursor's data to objects. An example was obj in qs, which only
converted the needed amount of objects. However, this case isn't the
most common one, and making everybody pay a performance penalty to
support this didn't seem like a good tradeoff.