Cleaned up join promotion logic in add_filter()
There was some logic to promote joins for ORed filters in add_filter().
The logic was somewhat complex, and what I can understand of it, it
wasn't correct.

The problem was that there was an assumed connection between
self.tables and the join list generated. The connection was that these
were iterated side by side. However, as far as I understand this kind
of connection isn't generally valid.

After this commit the code still passes all tests, but generates less
LEFT OUTER joins which should result in more efficient query plans.