Refactored save_base, part1
 - A change into saving logic - when saving an instance that came from
the same DB we are saving into use UPDATE - if not updated - INSERT
instead of SELECT - if found UPDATE - else INSERT. This should save a
query in majority of cases, for example in this case:
  a = SomeModel.objects.get(pk=pkval)
  # change a's fields
  a.save()
Previously a.save() would have done a SELECT - UPDATE, but the SELECT
is redundant as we know from model._state thate we just fetched the
model from the DB. In worst case the update doesn't actually update
anything and we will still correctly do an INSERT.

Also refactored the internal implementation of save_base() to be a
bit easier to follow. It should also allow customizing the save
behaviour in subclasses.