Generate join clause in field.get_join_sql()
The idea is to allow for completely custom join clauses, where
multicolumn joins are an interesting special case.

The main portion of the patch removes usage of alias_map column info
from various places of the ORM. There are still some references to
join columns, but it is not possible to create multicolumn joins in
the compiler.