Fixed #19385 -- Introduced multicolumn joins to the ORM
This patch consist of two main parts. The first part deals with adding
RelatedField and ForeignObject fields and basing current relation
fields on those. The second part deals with altering the ORM to be able
to handle multicolumn joins.

There is still a lot of work to be done for full multicolumn ForeignKey
support. The ForeignObject based fields do not integrate to other parts
of Django (admin, ModelForms, ...). There is no support for multicolumn
primary keys or any public way to use multicolumn ForeignKeys. Even
with these limitations the ForeignObject based fields allows
implementing some interesting constructs, for examples see the added
tests in tests/foreign_object/tests.py).

The downside of using ForeignObject is that it is private API and should
remain so until full virtual field support is implemented. Changes are
expected.

The patch was written by Jeremy Tillman with some final polish by
committer. A very big thank you for the work done!